A script for generating microservice projects
This aims to generate microservice projects in a consistent manner through cli. project structure is declared in a file called "API" in the directory this is being run, and follows the format "FRONT\nBACK\nROUTES\n-endpoint1\n--post\n--get\n\n-endpoint2\n--get\n--post\n--put\n--delete\n\n" etc. that is, verbs for a service fall beneath their respective endpoints, declared underneath ROUTES, and signified with a single "-", where verbs are signified via two "-". This is to streamline the rest of the microservice oriented architecture I am building.  NOTE the orchestrator is the microservice coordinator who hooks them together