The message component is minimally functional
Socketio handles interfacing between front and backend, and for group. The design incorporates all chats interfacing with the server, regardless of dm or group message designation.  In this sense, all chats are group messages, where sockets broadcast the signal and the server uses this as a signal to log for chat retrieval. React was difficult to get working with this, but state is handled properly now, with backend maintaining separate flags for 1) pending message 2) sent message.